% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addNormalizedAssays}
\alias{addNormalizedAssays}
\title{addNormalizedAssays}
\usage{
addNormalizedAssays(se, readLength = 50L)
}
\arguments{
\item{se}{A bin-wise `SummarizedExperiment` as produced by 
\code{\link{countFeatures}}}

\item{readLength}{Used as a minimum width to estimate read density 
(default 50).}
}
\value{
The `se` object with populated `logcpm` and `logNormDensity` assays.
}
\description{
addNormalizedAssays
}
\examples{
data(example_bin_se)
example_bin_se <- addNormalizedAssays(example_bin_se)
}
