% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countFeatures.R
\name{countFeatures}
\alias{countFeatures}
\title{countFeatures}
\usage{
countFeatures(
  bamfiles,
  bins,
  strandSpecific = 1,
  allowMultiOverlap = TRUE,
  inclNormalized = TRUE,
  ...
)
}
\arguments{
\item{bamfiles}{A vector of paths to bam files}

\item{bins}{A GRanges of bins in which to count reads (or path to a rds file
containing such an object}

\item{strandSpecific}{Passed to `Rsubread::featureCounts`}

\item{allowMultiOverlap}{Passed to `Rsubread::featureCounts`}

\item{inclNormalized}{Logical; whether to include normalized assays (needed for
plotting)}

\item{...}{Passed to `Rsubread::featureCounts`}
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment-class}
}
\description{
countFeatures
}
