% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{deuBinPlot}
\alias{deuBinPlot}
\title{deuBinPlot}
\usage{
deuBinPlot(
  se,
  gene,
  type = c("summary", "condition", "sample"),
  intronSize = 1,
  exonSize = c("sqrt", "linear", "log"),
  y = NULL,
  condition = NULL,
  size = "meanLogDensity",
  lineSize = 1,
  colour = NULL
)
}
\arguments{
\item{se}{A bin-wise SummarizedExperiment as produced by \code{\link{countFeatures}}
and including bin-level tests (i.e. having been passed through one of the DEU wrappers
such as \code{\link{diffSplice.wrapper}} or \code{\link{DEXSeq.wrapper}})}

\item{gene}{The gene of interest}

\item{type}{Either 'summary' (plot DEU summary), 'sample' (plot sample-wise data), or
'condition' (plot data aggregate by condition)}

\item{intronSize}{Intron plot size. If <=1, total intron size will be this fraction of
the total exon size. If >1, each intron will have the given size.}

\item{exonSize}{Scaling for exon sizes, either 'sqrt', 'log', or 'linear'.}

\item{y}{Value to plot on the y-axis. If `type="summary"`, this should be a column of
`rowData(se)`, otherwise should be an assay name of `se`.}

\item{condition}{The colData column containing the samples' condition.}

\item{size}{rowData variable to use to determine the thickness of the bins.}

\item{lineSize}{Size of the line connecting the bins. Use `lineSize=0` to omit the line.}

\item{colour}{rowData variable to use to determine the colour of the bins (can also
be `condition` or `sample` when using the corresponding plot `type`)}
}
\value{
A ggplot object
}
\description{
deuBinPlot
}
