% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_bin_se}
\alias{example_bin_se}
\title{Example bin-level `RangedSummarizedExperiment`}
\value{
a `RangedSummarizedExperiment`
}
\description{
An object produced by \code{\link{countFeatures}} containing small subset of
 genes from mouse hippocampal slices undergoing Forskolin-induced long-term
potentiation (GSE84643).
}
\references{
\url{https://www.nature.com/articles/s41598-017-17407-w}
}
