% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geneBinHeatmap}
\alias{geneBinHeatmap}
\title{geneBinHeatmap}
\usage{
geneBinHeatmap(
  se,
  gene,
  what = c("logNormDensity", "logCPM", "scaledLogCPM"),
  anno_rows = c("logWidth", "meanLogCPM", "meanLogDensity", "log10PValue"),
  ...
)
}
\arguments{
\item{se}{A bin-wise SummarizedExperiment as produced by \code{\link{countFeatures}}}

\item{gene}{The gene of interest}

\item{what}{Type of values (i.e. assay) to plot}

\item{anno_rows}{Row annotation columns (i.e. columns of `rowData(se)`) to plot}

\item{...}{Passed to `sechm` (see \code{\link[SEtools]{SE-heatmap}}).}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap}}
}
\description{
geneBinHeatmap
}
