% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rn6_PAS}
\alias{rn6_PAS}
\title{Poly-A sites compendium for Rattus Norvegicus (Rno6)}
\value{
a `GRanges` object
}
\description{
These are the sites from polyA_DB release 3.2, downloaded from
\url{https://exon.apps.wistar.org/PolyA_DB/v3/download/3.2/rat_pas.zip},
and lifted over to Rno6.
}
