% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLevel.R
\name{simes.aggregation}
\alias{simes.aggregation}
\title{simes.aggregation}
\usage{
simes.aggregation(p.value, geneid)
}
\arguments{
\item{p.value}{A vector of p-values}

\item{geneid}{A vector of group labels such as gene identifiers}
}
\value{
A named vector of aggregated p-values
}
\description{
Simes p-value correction and aggregation, adapted from \code{link[limma]{diffSplice}}
}
\examples{
p <- runif(50)
genes <- sample(LETTERS,50,replace=TRUE)
simes.aggregation(p, genes)
}
