% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addNormalizedAssays}
\alias{addNormalizedAssays}
\title{addNormalizedAssays}
\usage{
addNormalizedAssays(se)
}
\arguments{
\item{se}{A bin-wise SummarizedExperiment as produced by \code{\link{countFeatures}}}
}
\value{
The `se` object with populated `logcpm` and `logNormDensity` assays.
}
\description{
addNormalizedAssays
}
