% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geneBinHeatmap}
\alias{geneBinHeatmap}
\title{geneBinHeatmap}
\usage{
geneBinHeatmap(
  se,
  gene,
  what = c("logNormDensity", "logCPM", "scaledLogCPM"),
  anno_rows = c("type", "logWidth", "meanLogDensity", "log10PValue", "geneAmbiguous"),
  anno_colors = list(),
  removeAmbiguous = FALSE,
  merge_legends = TRUE,
  ...
)
}
\arguments{
\item{se}{A bin-wise SummarizedExperiment as produced by \code{\link{countFeatures}}}

\item{gene}{The gene of interest}

\item{what}{Type of values (i.e. assay) to plot}

\item{anno_rows}{Row annotation columns (i.e. columns of `rowData(se)`) to plot}

\item{anno_colors}{Annotation colors (passed to `SEtools::sechm`)}

\item{removeAmbiguous}{Logical; whether to remove bins that are
gene-ambiguous (i.e. overlap multiple genes).}

\item{merge_legends}{Logical; whether to merge legends. This effectively
calls `draw(..., merge_legends=TRUE)` around the heatmap.}

\item{...}{Passed to `sechm` (see \code{\link[SEtools]{SE-heatmap}}).}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap}}
}
\description{
geneBinHeatmap
}
\examples{
data(example_bin_se)
se <- diffSplice.wrapper(example_bin_se, ~condition)
geneBinHeatmap(se, "Jund")
}
